Added CMakeLists file. Created separate folder for this importer, with included main.cpp file as a working example
